DDCNLS-110: Converted SandboxFractionsCalculationControllerISpec.scala to FractionsCalculationDateEndpointISpec.scala funspec
