+++ it/uk/gov/hmrc/apprenticeshiplevy/IntegrationTestSuite.scala
@@ -8,7 +8,8 @@ class IntegrationTestsSuite extends Suites(new ServiceLocatorRegistrationISpec,
                                            new uk.gov.hmrc.apprenticeshiplevy.sandbox.Suite,
                                            new DocumentationEndpointISpec,
                                            new DefinitionEndpointISpec,
-                                           new EmploymentCheckEndpointISpec)
+                                           new EmploymentCheckEndpointISpec,
+                                           new FractionsCalculationDateEndpointISpec)
   with BeforeAndAfterAllConfigMap with IntegrationTestConfig {
 
   override def beforeAll(cm: ConfigMap) {
