+++ it/uk/gov/hmrc/apprenticeshiplevy/sandbox/Suite.scala
@@ -4,5 +4,4 @@ import org.scalatest._
 
 @DoNotDiscover
 class Suite extends Suites(new SandboxRootControllerISpec,
-                           new SandboxEmprefControllerISpec,
-                           new SandboxFractionsCalculationControllerISpec)
+                           new SandboxEmprefControllerISpec)
